add root page skeleton for desktop
